w3c-reports/12092014/gtk/runned-tests/webgl/compressedTexSubImage2D-expected.txt

CONSOLE MESSAGE: WebGL: INVALID_ENUM: compressedTexSubImage2D: invalid format
CONSOLE MESSAGE: WebGL: INVALID_ENUM: compressedTexSubImage2D: invalid format

PASS compressedTexSubImage2D 
PASS Calling function () {
    gl.compressedTexSubImage2D(gl.TEXTURE_2D, 0, 0, 0, 10, 10, COMPRESSED_RGB_S3TC_DXT1_EXT, new Uint8Array(8));
  } should generate a 1280 error. 
PASS Calling function () {
    gl.compressedTexSubImage2D(gl.TEXTURE_2D, 0, 0, 0, 10, 10, COMPRESSED_RGB_S3TC_DXT1_EXT, new Uint8Array(8), null);
  } should generate a 1280 error. 
PASS Should throw a TypeError when passing too few arguments. 
